refactor(ops): share Grafana inventory fixtures
