+++ scripts/ops/grafana-client.inventory.normalization.test.mjs
@@ -1,61 +1,24 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { readMetadataInventoryInternal } from "./grafana-metadata-inventory.mjs";
 import {
   INVENTORY_TARGET,
   completePlan,
   controlledRuntime,
+  datasource,
   emptyPlan,
-  fixtureFetch,
-  internalPrepare,
+  isDatasourceHealthRequest,
+  policyChain,
   rawResponse,
+  runControlledBody,
+  runPlan,
+  searchItem,
+  serviceAccount,
+  setSearchPages,
+  setServicePages,
   surfaceFailure,
   twoPageCompletePlan,
 } from "./grafana-client.inventory.fixtures.mjs";
-
-async function runPlan(plan = emptyPlan(), runtime = controlledRuntime()) {
-  const trace = [];
-  const fetchImpl = fixtureFetch(plan, { trace });
-  const result = await readMetadataInventoryInternal({ target: INVENTORY_TARGET,
-    prepareAuthorizedGet: internalPrepare(fetchImpl), productionClockAndTimers: runtime.dependencies });
-  return { result, trace, runtime };
-}
-
-function datasource(uid, overrides = {}) {
-  return { uid, name: `name-${uid}`, type: "cloudwatch", access: "proxy",
-    isDefault: false, readOnly: true, ...overrides };
-}
-
-function searchItem(type, uid, overrides = {}) {
-  return { type, uid, title: `title-${uid}`, ...overrides };
-}
-
-function serviceAccount(id, overrides = {}) {
-  return { id, name: `name-${id}`, role: "Viewer", isDisabled: false, tokens: 0, ...overrides };
-}
-
-function setSearchPages(plan, type, pages) {
-  for (let index = 0; index < pages.length; index += 1) {
-    plan.set(`/api/search?type=${type}&limit=100&page=${index + 1}`, pages[index]);
-  }
-}
-
-function setServicePages(plan, pages) {
-  for (let index = 0; index < pages.length; index += 1) {
-    plan.set(`/api/serviceaccounts/search?perpage=100&page=${index + 1}`, pages[index]);
-  }
-}
-
-function isDatasourceHealthRequest({ path }) {
-  return path.startsWith("/api/datasources/uid/") && path.endsWith("/health");
-}
-
-function policyChain(depth, leaf = { receiver: "leaf" }) {
-  let node = leaf;
-  for (let index = 0; index < depth; index += 1) node = { receiver: `r${index}`, routes: [node] };
-  return node;
-}
 
 const PRECEDENCE_CONTRACT = Object.freeze({
   envelope: [1, "malformed", "invalid_shape"], apiState: [2, "unavailable", "api_database_not_ok"],
@@ -494,23 +457,6 @@ test("Unicode, path, number, duplicate, and string limits fail closed before fan
       surfaceFailure("malformed", "invalid_shape", 200));
   }
 });
-
-async function runControlledBody(path, body, expiresAtObservation, plan = emptyPlan()) {
-  const runtime = controlledRuntime();
-  let armed = false;
-  let observations = 0;
-  runtime.dependencies.monotonicNow = () => {
-    if (!armed) return 0;
-    observations += 1;
-    return observations >= expiresAtObservation ? 30_000 : 0;
-  };
-  plan.set(path, rawResponse({
-    body,
-    onRead(index) { if (index === 1) armed = true; },
-  }));
-  const { result, trace } = await runPlan(plan, runtime);
-  return { result, trace, observations };
-}
 
 test("Ruler row 4 and row 6 are atomic, order-independent, and precede invalid keys", async (t) => {
   const invalidNamespace = "\ud800";
